functional tests: Factor out some common code
