+++ test/functional/p2p_relay_transactions.py
@@ -4,7 +4,7 @@
 # Distributed under the MIT software license, see the accompanying
 # file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-from test_framework import mintlayer_hash
+from test_framework.mintlayer import mintlayer_hash
 from test_framework.util import assert_raises_rpc_error
 from test_framework.test_framework import BitcoinTestFramework
 import scalecodec
